requirements.txt
google-genai
tavily-python
python-dotenv
streamlit
pandas
plotly
requests
